+++ internal/app/http/service.go
@@ -163,7 +163,7 @@ func (dhs *defaultHandlerService) ImportThingModel(ctx context.Context, file []b
 
 	repo, err := repos.Get(importRepo)
 	if err != nil {
-		return repos.ImportResult{}, err
+		return repos.ImportResultFromError(err)
 	}
 	res, err := commands.NewImportCommand(time.Now).ImportFile(ctx, file, repo, opts)
 	if err != nil {
@@ -172,7 +172,7 @@ func (dhs *defaultHandlerService) ImportThingModel(ctx context.Context, file []b
 	if res.IsSuccessful() {
 		err = repo.Index(ctx, res.TmID)
 		if err != nil {
-			return repos.ImportResult{}, err
+			return repos.ImportResultFromError(err)
 		}
 	}