+++ internal/repos/tmc.go
@@ -216,16 +216,16 @@ func (t TmcRepo) Import(ctx context.Context, id model.TMID, raw []byte, opts Imp
 	reqUrl.RawQuery = vals.Encode()
 	req, err := http.NewRequestWithContext(ctx, http.MethodPost, reqUrl.String(), bytes.NewBuffer(raw))
 	if err != nil {
-		return ImportResult{}, err
+		return ImportResultFromError(err)
 	}
 	req.Header.Add(headerContentType, mimeJSON)
 	resp, err := doHttp(req, t.auth)
 	if err != nil {
-		return ImportResult{}, err
+		return ImportResultFromError(err)
 	}
 	b, err := io.ReadAll(resp.Body)
 	if err != nil {
-		return ImportResult{}, err
+		return ImportResultFromError(err)
 	}
 
 	switch resp.StatusCode {
@@ -233,7 +233,7 @@ func (t TmcRepo) Import(ctx context.Context, id model.TMID, raw []byte, opts Imp
 		var res server.ImportThingModelResponse
 		err = json.Unmarshal(b, &res)
 		if err != nil {
-			return ImportResult{}, err
+			return ImportResultFromError(err)
 		}
 		msg := ""
 		if res.Data.Message != nil {
@@ -242,7 +242,7 @@ func (t TmcRepo) Import(ctx context.Context, id model.TMID, raw []byte, opts Imp
 		if res.Data.Code != nil && *res.Data.Code != "" {
 			cErr, err := ParseErrTMIDConflict(*res.Data.Code)
 			if err != nil {
-				return ImportResult{}, err
+				return ImportResultFromError(err)
 			}
 			return ImportResult{Type: ImportResultWarning, TmID: res.Data.TmID, Message: msg, Err: cErr}, nil
 		}
@@ -251,7 +251,7 @@ func (t TmcRepo) Import(ctx context.Context, id model.TMID, raw []byte, opts Imp
 		var e server.ErrorResponse
 		err = json.Unmarshal(b, &e)
 		if err != nil {
-			return ImportResult{}, err
+			return ImportResultFromError(err)
 		}
 		detail := e.Title
 		if e.Detail != nil {
@@ -265,19 +265,23 @@ func (t TmcRepo) Import(ctx context.Context, id model.TMID, raw []byte, opts Imp
 			}
 			cErr, err := ParseErrTMIDConflict(eCode)
 			if err != nil {
-				return ImportResult{}, err
+				return ImportResultFromError(err)
 			}
-			return ImportResult{}, cErr
+			return ImportResult{
+				Type:    ImportResultTMExists,
+				Message: cErr.Error(),
+				Err:     cErr,
+			}, cErr
 		case http.StatusInternalServerError, http.StatusUnauthorized, http.StatusBadRequest:
 			err := errors.New(detail)
-			return ImportResult{}, err
+			return ImportResultFromError(err)
 		default:
 			err := errors.New("unexpected status not handled correctly")
-			return ImportResult{}, err
+			return ImportResultFromError(err)
 		}
 	default:
 		err := errors.New(fmt.Sprintf("received unexpected HTTP response from remote TM catalog: %s", resp.Status))
-		return ImportResult{}, err
+		return ImportResultFromError(err)
 	}
 }
 func (t TmcRepo) Delete(ctx context.Context, id string) error {